アラート表示のプロトコルエクステンション作成
